% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_calibration.R
\name{apply_shift}
\alias{apply_shift}
\title{Apply spatial shift to GEDI data}
\usage{
apply_shift(gedi_obs, shift)
}
\arguments{
\item{gedi_obs}{Original GEDI observations}

\item{shift}{shift containing one unique shift per file. Preferably output of \code{\link[=get_best_shift]{get_best_shift()}}}
}
\value{
Calibrated GEDI observations
}
\description{
Apply spatial shift to GEDI data
}
\examples{
library(sf)
gedi = system.file("demodata/gedi.gpkg", package = "GEDIcalibratoR") |> st_read()
grid_ext = system.file("demodata/grid_ext.rds", package = "GEDIcalibratoR") |> readRDS()
shifts = shift_stats(grid_ext)
best_shift = get_best_shift(grid_ext, shifts)
gedi_cal = apply_shift(gedi, best_shift)
}
