% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_calibration.R
\name{cal_extract}
\alias{cal_extract}
\title{Extract reference raster values for a calibration grid}
\usage{
cal_extract(tile_paths, gedi_grid, buffer = 12.5, fun = "max", quiet = F)
}
\arguments{
\item{tile_paths}{Paths to DSM, DEM or preprocessed RHM tiles.
Output of \code{\link[=download_tiles]{download_tiles()}} or \code{\link[=make_rhm]{make_rhm()}}.}

\item{gedi_grid}{Calibration grid. Output of \code{\link[=cal_grid_rect]{cal_grid_rect()}} or \code{\link[=cal_grid_circle]{cal_grid_circle()}}.}

\item{buffer}{spatial circular buffer around each grid point. Default is 12.5 m to match GEDI's 25 m footprint.}

\item{fun}{see \code{\link[terra:extract]{terra::extract()}}}

\item{quiet}{If FALSE (default), print progress messages.}
}
\value{
The same input calibration grid (sf object) with an additional column named 'ext_value'.
}
\description{
Extract reference raster values for a calibration grid
}
\examples{
\dontrun{
data(gedi)
gedi_circ = cal_grid_circle(gedi)
tiles = intersect_tiles2download(gedi_obs = gedi_circ)
down_tiles = download_tiles(tiles, dir = tempdir())
rhm_paths = make_rhm(tile_paths, overwrite_if_exists = T)

grid_ext = cal_extract(rhm_paths, gedi_circ, buffer = 12.5)
summary(grid_ext)
hist(grid_ext$ext_value)
}
}
