% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_calibration.R
\name{cal_grid_circle}
\alias{cal_grid_circle}
\alias{cal_grid_rect}
\title{Create grid layouts for GEDI location calibration}
\usage{
cal_grid_circle(
  gedi,
  steps = c(0, 5, 10),
  n_circle_elements = 8,
  copy_GEDI_vars = NULL
)

cal_grid_rect(gedi, steps = c(0, 5, 10), copy_GEDI_vars = NULL)
}
\arguments{
\item{gedi}{GEDI point locations}

\item{steps}{Sequence (vector) of sample point distances from the original GEDI location to span the grid. Should start with 0.}

\item{n_circle_elements}{Number of points per circle. Only applies to cal_grid_circle(). Defaults to 8.}

\item{copy_GEDI_vars}{Which GEDI variables to keep for calibration, e.g. "rh98".}
}
\value{
sf object containing calibration grids for each GEDI observation.
}
\description{
\code{\link[=cal_grid_rect]{cal_grid_rect()}} and \code{\link[=cal_grid_circle]{cal_grid_circle()}} both create grids to be used in the location
calibration process. The original GEDI observations serves as the center of the grid. GEDI Level 2 data comes with a
mean location error of 10 meters. Its relationship with ALS data may be enhanced through calibration.
}
\examples{
library(sf)
library(dplyr)
gedi = system.file("demodata/gedi.gpkg", package = "GEDIcalibratoR") |> st_read()

gedi_rect = cal_grid_rect(gedi, steps = c(0, 3, 6, 9, 12),
                          copy_GEDI_vars = c("file","id","rh98"))
gedi_circ = cal_grid_circle(gedi, steps = c(0, 3, 6, 9, 12),
                            n_circle_elements = 12,
                            copy_GEDI_vars = c("file","id","rh98"))
par(mfrow=c(1,2))
filter(gedi_rect) |> slice_head(n = 1) |> st_geometry() |> plot()
filter(gedi_circ) |> slice_head(n = 1) |> st_geometry() |> plot()
}
