% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiles_download.R
\name{download_als_zip}
\alias{download_als_zip}
\alias{download_als_gz}
\alias{download_als_laz}
\title{Download LAZ, DEM, and DSM tiles}
\usage{
download_als_zip(tile, dir, tmp)

download_als_gz(tile, dir, tmp)

download_als_laz(tile, dir)
}
\arguments{
\item{tile}{single row from tiles table, an output of \code{\link[=intersect_tiles2download]{intersect_tiles2download()}}
or \code{\link[=select_tiles2download]{select_tiles2download()}}.}

\item{dir}{directory to download tiles to.}

\item{tmp}{temporary file}
}
\description{
Internal function used by \code{\link[=download_tiles]{download_tiles()}}
}
