% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEDIcalibratoR-package.R
\docType{package}
\name{GEDIcalibratoR-package}
\alias{GEDIcalibratoR}
\alias{GEDIcalibratoR-package}
\title{GEDIcalibratoR: Location Calibration For GEDI Using ALS Reference Data}
\description{
GEDIcalibratoR facilitates location calibration for GEDI. It makes downloading tiles of ALS, DEM, and DEM data easy and evaluates possible spatial shifts using calibration grids. The best suitable shift can then be applied to the original data and may improve statistical models and following analyses.
}
\author{
\strong{Maintainer}: Johannes Heisig \email{jheisig@uni-muenster.de} (\href{https://orcid.org/0000-0003-3586-3001}{ORCID})

}
\keyword{internal}
