% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiles_download.R
\name{intersect_tiles2download}
\alias{intersect_tiles2download}
\title{Get ALS tiles by intersection}
\usage{
intersect_tiles2download(region = "all", gedi_obs = NULL, quiet = F)
}
\arguments{
\item{region}{Region to select tiles from. Options are 'NRW', 'TH', 'SN', or 'all' (default).}

\item{gedi_obs}{Point locations of GEDI observations. Should be a \code{sf} or \code{sfc} object.}

\item{quiet}{If TRUE, hide message with ALS tile names. Default is FALSE.}
}
\value{
sf object with download links for tiles of ALS point clouds,
digital elevation models, and digital surface models. Used as input to download_tiles().
}
\description{
Non-interactive version of \code{select_tiles2download()}. Takes an input of point locations and finds intersecting ALS tiles.
Currently supports open LiDAR archives from three German states: Northrhine-Westfalia (NRW), Thuringia (TH), and Saxony (SN).
}
\examples{
print("hi")
}
\seealso{
\code{\link[=select_tiles2download]{select_tiles2download()}}
}
