% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_calibration.R
\name{make_rhm}
\alias{make_rhm}
\title{Create Relative Height Model from DEM and DSM}
\usage{
make_rhm(
  tile_paths,
  aggregate_res = NULL,
  overwrite_if_exists = F,
  quiet = F,
  ...
)
}
\arguments{
\item{tile_paths}{Paths to downloaded DSM and DEM tile. Output of \code{\link[=download_tiles]{download_tiles()}}.}

\item{aggregate_res}{Aggregation factor to decrease spatial detail of RHM.}

\item{overwrite_if_exists}{Recompute and overwrite existing RHM tiles?}

\item{quiet}{If TRUE, hide message with tile names. Default is FALSE.}

\item{...}{Further arguments to terra::aggregate(), e.g. fun = "max"}
}
\value{
paths to RHM tiles
}
\description{
A Relative Height Model (RHM) is a normalized Digital Surface Model (DSM). It is created by simply deducting the
Digital Elevation Model (DEM) from the DSM.
}
\examples{
\dontrun{
data(gedi)
tiles = intersect_tiles2download(gedi_obs = gedi)
down_tiles = download_tiles(tiles, dir = tempdir())
rhm_paths = make_rhm(tile_paths, overwrite_if_exists = T)
}
}
