% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_calibration.R
\name{merge_tiles}
\alias{merge_tiles}
\title{Merge downloaded raster tiles}
\usage{
merge_tiles(paths, CRS)
}
\arguments{
\item{paths}{paths downloaded raster tiles}

\item{CRS}{force assign CRS, e.g. "EPSG:25832"}
}
\value{
SpatRaster object from merging one or multiple tiles
}
\description{
Merge downloaded raster tiles
}
