% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_calibration.R
\name{plot_calibration}
\alias{plot_calibration}
\title{Plot calibration results}
\usage{
plot_calibration(cal, ref = NULL, label = "", color = NULL)
}
\arguments{
\item{cal}{Calibrated GEDI locations. Output of \code{\link[=get_best_shift]{get_best_shift()}}.}

\item{ref}{Optional. Reference GEDI locations, e.g. output of \code{\link[=get_no_shift]{get_no_shift()}}.}

\item{label}{Length 2 character vector. Name of the table to be displayed. e.g. "best shift" and "reference".}

\item{color}{Length 2 character vector. Colors identifying ref and cal.}
}
\description{
Plot calibration results
}
\examples{
grid_ext = system.file("demodata/grid_ext.rds", package = "GEDIcalibratoR") |> readRDS()
shifts = shift_stats(grid_ext)
bs = get_best_shift(grid_ext, shifts)
ns = get_no_shift(grid_ext, shifts)
plot_calibration(bs, ns,  label = c("Best","Ref"), color =  c("forestgreen", "firebrick"))
}
