% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_calibration.R
\name{plot_shift_stats_rect}
\alias{plot_shift_stats_rect}
\alias{plot_shift_stats_polar}
\title{Comparative scatter plot of calibration shifts}
\usage{
plot_shift_stats_rect(shifts)

plot_shift_stats_polar(shifts)
}
\arguments{
\item{shifts}{Output of \code{\link[=shift_stats]{shift_stats()}}}
}
\description{
Comparative scatter plot of calibration shifts
}
