% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_calibration.R
\name{rmse}
\alias{rmse}
\alias{r2}
\title{R² and RMSE helper functions}
\usage{
rmse(a, p)

r2(a, p)
}
\arguments{
\item{a}{actual values}

\item{p}{predicted values}
}
\value{
numeric
}
\description{
R² and RMSE helper functions
}
