% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_calibration.R
\name{shift_stats}
\alias{shift_stats}
\title{Calculate statistics for each calibration shift}
\usage{
shift_stats(ext_data, cal_var = "rh98")
}
\arguments{
\item{ext_data}{Extracted reference values. Output of \code{\link[GEDIcalibrator:cal_extract]{GEDIcalibrator::cal_extract()}}.}

\item{cal_var}{GEDI variable in ext_data to use for comparison.}
}
\value{
Table with an entry for each file and shift summarizing calibratiion measures RMSE and R².
}
\description{
Calculate statistics for each calibration shift
}
