% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiles_download.R
\name{download_tiles}
\alias{download_tiles}
\title{Download selected LAZ, DEM, and DSM tiles}
\usage{
download_tiles(tiles, dir, what = "DSM", setTimeOut = NULL)
}
\arguments{
\item{tiles}{table containing tiles to donload. output of \code{\link[=intersect_tiles2download]{intersect_tiles2download()}}
or \code{\link[=select_tiles2download]{select_tiles2download()}}.}

\item{dir}{directory to download tiles to.}

\item{what}{either "LAZ", "DEM", or "DSM" (default).}

\item{setTimeOut}{If e.g. laz files are large, download performance may suffer from the default
timeout setting (60 sec). Adjust this in case you run into difficulties.}
}
\value{
paths to downloaded files
}
\description{
Download selected LAZ, DEM, and DSM tiles
}
\examples{
\dontrun{
data("gedi")
tiles = intersect_tiles2download(gedi_obs=gedi)

# subset to keep one tile from each region
library(dplyr)
tiles = group_by(tiles, state) |> slice_head(n=1)

# download to temporary directory
td = tempdir()
dsm_tiles = download_tiles(tiles, dir = td, what = "DSM")
dem_tiles = download_tiles(tiles, dir = td, what = "DEM")
laz_tiles = download_tiles(tiles, dir = td, what = "LAZ", setTimeOut = 300)

file.exists(c(dsm_tiles, dem_tiles, laz_tiles))

# delete files
unlink(c(dsm_tiles, dem_tiles, laz_tiles))
}
}
